fix: moved shared modules into shared scope
